Refactorize host and hostgroup models for readability sake
