+++ models/hostgroup/public_functions.rb
@@ -1,18 +1,5 @@
 class Hostgroup
-  include DataMapper::Resource
   include Monitoring::Hostgroup
-
-  def self.default_repository_name #here we use the hosts_db for the Hostgroup objects
-   :hosts_db
-  end
-
-  property :name, String, :key => true
-  property :autodeploy, Object
-  property :opt_vars, Object, :default  => {}
-  property :created_at, DateTime
-  property :updated_at, DateTime
-  has n, :hostgroup_members
-  has n, :hosts, :through => :hostgroup_members
 
   def add_member(host)
     self.hosts << host
@@ -81,15 +68,5 @@ class Hostgroup
     end
     Process.waitall
   end
-end
 
-class HostgroupMember
-  include DataMapper::Resource
-
-  def self.default_repository_name #here we use the hosts_db for the HostgroupMember relation
-   :hosts_db
-  end
-
-  belongs_to :host,   :key => true
-  belongs_to :hostgroup, :key => true
 end
